requirements.txt
supabase
python-dotenv
requests
httpx
outscraper
serpapi
apify-client
